requirements.txt
requests
rich
pyside6
mysql-connector-python
fastapi
uvicorn[standard]
